# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/wallet": {
      "get": {
        "operationId": "WalletController_me",
        "summary": "My wallet balance + default leverage (contract §wallet)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              }
            }
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/wallet/ledger": {
      "get": {
        "operationId": "WalletController_ledger",
        "summary": "My wallet ledger — recent money movements",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerEntryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/wallet/transactions": {
      "get": {
        "operationId": "WalletController_transactions_",
        "summary": "My unified transaction timeline (deposits, trades, fees, withdrawals)",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserTxnDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/wallet/payout-methods": {
      "get": {
        "operationId": "PayoutMethodsController_list",
        "summary": "My saved withdrawal methods (masked)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PayoutMethodDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PayoutMethodsController_create",
        "summary": "Add a withdrawal method (crypto / bank; encrypted)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayoutMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutMethodDto"
                }
              }
            }
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/wallet/payout-methods/{id}": {
      "delete": {
        "operationId": "PayoutMethodsController_remove",
        "summary": "Remove a withdrawal method",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "wallet"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/youtube/status": {
      "get": {
        "operationId": "YoutubeController_status",
        "summary": "Is my YouTube account connected?",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "youtube"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/youtube/connect": {
      "post": {
        "operationId": "YoutubeController_connect",
        "summary": "Get the Google consent URL to connect YouTube",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "youtube"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/youtube/disconnect": {
      "post": {
        "operationId": "YoutubeController_disconnect",
        "summary": "Disconnect my YouTube account",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "youtube"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/broadcasts/{id}/destinations/youtube/connect": {
      "post": {
        "operationId": "BroadcastsController_connectYoutube",
        "summary": "Begin YouTube simulcast OAuth (contract §4.9)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "broadcasts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Millimore API",
    "description": "Backend API for the Millimore platform (mobile app + admin dashboard). Single source of truth: docs/BACKEND_CONTRACT.md",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.millimore.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WalletDto": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "example": 1250.75,
            "description": "Spendable balance = free margin."
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "leverage": {
            "type": "number",
            "example": 100,
            "description": "Default leverage (1..maxLeverage)."
          }
        },
        "required": [
          "balance",
          "currency",
          "leverage"
        ]
      },
      "LedgerEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "led_ab12cd34"
          },
          "type": {
            "type": "string",
            "example": "deposit",
            "description": "deposit | withdrawal_hold | withdrawal_refund | copy_allocate | copy_return | trade_pnl | commission | commission_earned | platform_fee | adjustment"
          },
          "amount": {
            "type": "number",
            "example": 250,
            "description": "Signed: credits positive, debits negative."
          },
          "balanceAfter": {
            "type": "number",
            "example": 1250.75
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "refId": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-02T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "type",
          "amount",
          "balanceAfter",
          "currency",
          "createdAt"
        ]
      },
      "UserTxnDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "led_ab12cd34"
          },
          "kind": {
            "type": "string",
            "example": "deposit",
            "description": "deposit | withdrawal | trade | commission | fee | transfer"
          },
          "title": {
            "type": "string",
            "example": "Deposit USDT"
          },
          "amount": {
            "type": "number",
            "example": 250,
            "description": "Signed: credits positive, debits negative."
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "status": {
            "type": "string",
            "example": "completed",
            "description": "completed | pending | rejected"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-02T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "amount",
          "currency",
          "status",
          "createdAt"
        ]
      },
...
      "PayoutMethodDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pm_ab12cd34"
          },
          "type": {
            "type": "string",
            "enum": [
              "crypto",
              "bank"
            ]
          },
          "label": {
            "type": "string",
            "example": "My USDT wallet"
          },
          "masked": {
            "type": "string",
            "example": "USDT ••••a1b2"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-02T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "type",
          "label",
          "masked",
          "createdAt"
        ]
      },
      "CreatePayoutMethodDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "crypto",
              "bank"
            ]
          },
          "label": {
            "type": "string",
            "example": "My USDT wallet"
          },
          "asset": {
            "type": "string",
            "example": "USDT"
          },
          "address": {
            "type": "string",
            "example": "TXyz…addr"
          },
          "bankName": {
            "type": "string",
            "example": "Emirates NBD"
          },
          "accountNumber": {
            "type": "string",
            "example": "00012345678"
          },
          "iban": {
            "type": "string",
            "example": "AE070331234567890123456"
          }
        },
        "required": [
          "type",
          "label"
        ]
      },
...
    }
  }
}